feat: sub01[02] func1 - add contact
