--- app/src/main/res/layout/activity_main.xml
@@ -1,189 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity">
-
-    <!--profile image-->
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="80dp"
-        android:layout_height="80dp"
-        android:layout_marginTop="32dp"
-        android:contentDescription="Profile Image"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:src="@drawable/profile_image"/>
-
-    <!--input name-->
-    <EditText
-        android:id="@+id/nameText"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:hint="이름"
-        android:layout_marginTop="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"  
-        app:layout_constraintTop_toBottomOf="@+id/imageView" 
-        android:inputType="text"  
-        android:padding="10dp" 
-        android:background="@drawable/round_background" /> 
-    
-    <!--input number-->
-    <EditText
-        android:id="@+id/phoneText"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:hint="전화번호"
-        android:layout_marginTop="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/nameText"
-        android:inputType="phone"
-        android:padding="10dp"
-        android:background="@drawable/round_background"/>
-    
-    <!--input email-->
-    <EditText
-        android:id = "@+id/emailText"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:hint="이메일"
-        android:layout_marginTop="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/phoneText"
-        android:inputType="textEmailAddress"
-        android:padding="10dp"
-        android:background="@drawable/round_background" />
-
-    <!--더보기-->
-    <TextView
-        android:id="@+id/moreText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="더보기"
-        android:textColor="@android:color/black"
-        android:padding="10dp"
-        android:layout_marginTop="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/emailText"
-        android:onClick="toggleMore" />
-
-    <!--더보기 아래 생일, 성별, 메모 레이아웃 묶기-->
-    <LinearLayout
-        android:id="@+id/moreLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:visibility="gone"
-        android:layout_marginTop="16dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/moreText">
-
-        <!--birth calendar-->
-        <TextView
-            android:id="@+id/birthText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="생일"
-            android:inputType="none"
-            android:padding="10dp"
-            android:background="@drawable/round_background"
-            android:onClick="showDatePicker" />
-
-        <!--choose gender layout-->
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:padding="10dp"
-            android:gravity="end"
-            android:layout_marginTop="16dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:background="@drawable/round_background">
-
-            <!--choose gender text-->
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="성별"
-                android:textColor="@android:color/darker_gray"
-                android:padding="10dp"
-                android:layout_gravity="center_vertical"/>
-
-            <!--real choose gender - radio-->
-            <RadioGroup
-                android:id="@+id/genderRadio"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal">
-
-                <!--female radio-->
-                <RadioButton
-                    android:id="@+id/femaleButton"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="여성"/>
-
-                <!--male radio-->
-                <RadioButton
-                    android:id="@+id/maleButton"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="남성"/>
-            </RadioGroup>
-        </LinearLayout>
-
-        <!--input memo-->
-        <EditText
-            android:id="@+id/memoText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="메모"
-            android:inputType="text"
-            android:layout_marginTop="16dp"
-            android:padding="10dp"
-            android:background="@drawable/round_background" />
-    </LinearLayout>
-
-    <!--ButtonLayout-->
-    <LinearLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <!--Cancel Button-->
-        <Button
-            android:id="@+id/cancelButton"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="취소"
-            android:textColor="@android:color/white"
-            android:layout_margin="8dp"
-            android:background="@drawable/round_background"/>
-
-        <!--Store Button-->
-        <Button
-            android:id="@+id/saveButton"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="저장"
-            android:textColor="@android:color/white"
-            android:layout_margin="8dp"
-            android:background="@drawable/round_background"/>
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
